removed preview toggles related code and move preview stuff from Work to IIIFPresentationPreview
