+++ catalogue/webapp/pages/work.js
@@ -1,5 +1,4 @@
 // @flow
-import { Fragment, useState, useEffect } from 'react';
 import Router from 'next/router';
 import NextLink from 'next/link';
 import fetch from 'isomorphic-unfetch';
@@ -26,113 +25,9 @@ import WorkDetails from '../components/WorkDetails/WorkDetails';
 import SearchForm from '../components/SearchForm/SearchForm';
 import TogglesContext from '@weco/common/views/components/TogglesContext/TogglesContext';
 import { getWork } from '../services/catalogue/works';
-// import IIIFPresentationPreview from '@weco/common/views/components/IIIFPresentationPreview/IIIFPresentationPreview';
+import IIIFPresentationPreview from '@weco/common/views/components/IIIFPresentationPreview/IIIFPresentationPreview';
 // import IIIFImagePreview from '@weco/common/views/components/IIIFImagePreview/IIIFImagePreview'; // TODO rename WorkMedia?, these are the same thing
 
-// TODO flow
-// TODO move to IIIFPresentationPreview
-// Ideal preview thumbnails order: Title page, Front Cover, first page of Table of Contents, 2 random.
-// If we don't have any of the sructured pages, we fill with random ones, so there are always 5 images if possible.
-function randomImages(iiifManifest = null, structuredImages = [], n = 1) {
-  const images = [];
-  const canvases = iiifManifest
-    ? iiifManifest.sequences
-        .find(sequence => sequence['@type'] === 'sc:Sequence')
-        .canvases.filter(canvas => {
-          // Don't include the structured pages we're using when getting random ones
-          return (
-            structuredImages
-              .map(type => {
-                return type.images.find(image => {
-                  return canvas['@id'] === image.canvasId;
-                });
-              })
-              .filter(Boolean).length === 0
-          );
-        })
-    : [];
-
-  const numberOfImages = canvases.length < n ? canvases.length : n;
-
-  for (var i = 1; i <= numberOfImages; i++) {
-    const randomNumber = Math.floor(Math.random() * canvases.length);
-    const randomCanvas = canvases.splice(randomNumber, 1)[0];
-    images.push({
-      orientation:
-        randomCanvas.thumbnail.service.width >
-        randomCanvas.thumbnail.service.height
-          ? 'landscape'
-          : 'portrait',
-      uri: iiifImageTemplate(randomCanvas.thumbnail.service['@id'])({
-        size: '!400,400',
-      }),
-      canvasId: randomCanvas['@id'],
-    });
-  }
-  return {
-    label: 'random',
-    images,
-  };
-}
-
-function structuredImages(iiifManifest = null) {
-  const structures = iiifManifest ? iiifManifest.structures : [];
-  return structures.map(structure => {
-    const images = structure.canvases.map(canvasId => {
-      const matchingCanvas = iiifManifest.sequences
-        .find(sequence => sequence['@type'] === 'sc:Sequence')
-        .canvases.find(canvas => {
-          return canvas['@id'] === canvasId;
-        });
-      return {
-        orientation:
-          matchingCanvas.thumbnail.service.width >
-          matchingCanvas.thumbnail.service.height
-            ? 'landscape'
-            : 'portrait',
-        uri: iiifImageTemplate(matchingCanvas.thumbnail.service['@id'])({
-          size: '!400,400',
-        }),
-        canvasId,
-      };
-    });
-    return {
-      label: structure.label,
-      images,
-    };
-  });
-}
-
-function orderedStructuredImages(structuredImages) {
-  const titlePage = structuredImages.find(
-    structuredImage => structuredImage.label === 'Title Page'
-  );
-  const frontCover = structuredImages.find(
-    structuredImage =>
-      structuredImage.label === 'Front Cover' ||
-      structuredImage.label === 'Cover'
-  );
-  const firstTableOfContents = structuredImages.find(
-    structuredImage => structuredImage.label === 'Table of Contents'
-  );
-  return [titlePage, frontCover, firstTableOfContents].filter(Boolean);
-}
-
-function previewThumbnails(
-  iiifManifest = {},
-  structuredImages = [],
-  idealNumber = 5
-) {
-  return structuredImages.length < idealNumber
-    ? structuredImages.concat(
-        randomImages(
-          iiifManifest,
-          structuredImages,
-          idealNumber - structuredImages.length
-        )
-      )
-    : structuredImages;
-}
 type Props = {|
   work: Work | CatalogueApiError,
   iiifManifest: ?{},
@@ -150,17 +45,6 @@ export const WorkPage = ({
   workType,
   itemsLocationsLocationType,
 }: Props) => {
-  const [imageThumbnails, setImageThumbnails] = useState(false);
-  useEffect(() => {
-    setImageThumbnails(
-      previewThumbnails(
-        iiifManifest,
-        orderedStructuredImages(structuredImages(iiifManifest)),
-        5
-      )
-    );
-  }, []);
-
   if (work.type === 'Error') {
     return (
       <ErrorPage
@@ -281,17 +165,6 @@ export const WorkPage = ({
         </div>
       </div>
 
-      {imageThumbnails &&
-        imageThumbnails.map(pageType => {
-          return pageType.images.map(image => (
-            <img
-              style={{ width: 'auto', maxHeight: '300px' }}
-              key={image.uri}
-              src={image.uri}
-            />
-          ));
-        })}
-
       <div
         className={classNames({
           row: true,
@@ -306,42 +179,41 @@ export const WorkPage = ({
       </div>
 
       {iiifManifest && (
-        <Layout12>
-          <NextLink
-            {...itemUrl({
-              workId: work.id,
-              query,
-              page,
-              workType,
-              itemsLocationsLocationType,
-            })}
-          >
-            <a>View item</a>
-          </NextLink>
-        </Layout12>
+        <>
+          <IIIFPresentationPreview manifestData={iiifManifest} />
+          <Layout12>
+            <NextLink
+              {...itemUrl({
+                workId: work.id,
+                query,
+                page,
+                workType,
+                itemsLocationsLocationType,
+              })}
+            >
+              <a>View item</a>
+            </NextLink>
+          </Layout12>
+        </>
       )}
 
-      <TogglesContext.Consumer>
-        <Fragment>
-          {iiifImageLocationUrl && (
-            <WorkMedia
-              id={work.id}
-              iiifUrl={iiifImageLocationUrl}
-              title={work.title}
-            />
-          )}
+      {iiifImageLocationUrl && (
+        <WorkMedia
+          id={work.id}
+          iiifUrl={iiifImageLocationUrl}
+          title={work.title}
+        />
+      )}
 
-          <WorkDetails
-            work={work}
-            iiifManifest={iiifManifest}
-            iiifImageLocationUrl={iiifImageLocationUrl}
-            licenseInfo={licenseInfo}
-            iiifImageLocationCredit={iiifImageLocationCredit}
-            iiifImageLocationLicenseId={iiifImageLocationLicenseId}
-            encoreLink={encoreLink}
-          />
-        </Fragment>
-      </TogglesContext.Consumer>
+      <WorkDetails
+        work={work}
+        iiifManifest={iiifManifest}
+        iiifImageLocationUrl={iiifImageLocationUrl}
+        licenseInfo={licenseInfo}
+        iiifImageLocationCredit={iiifImageLocationCredit}
+        iiifImageLocationLicenseId={iiifImageLocationLicenseId}
+        encoreLink={encoreLink}
+      />
     </PageLayout>
   );
 };
